Admin

+++ back/controllers/user.js
@@ -1,6 +1,5 @@
 const db = require("../models");
 const User = db.users;
-
 const bcrypt = require('bcrypt');
 const fs = require('fs');
 const jwt = require('jsonwebtoken');
@@ -83,7 +82,7 @@ exports.modifyUser = (req, res) => {
 
     User.findByPk(id)
         .then(user => {
-            if (user.id == userId) {
+            if (user.id == userId || user.isAdmin == 1) {
                 if (!req.body.password) {
                 if (req.file) {
                     image= `${req.protocol}://${req.get('host')}/images/${req.file.filename}`
@@ -129,34 +128,85 @@ exports.modifyUserPicture = (req, res) => {
 
     const id = req.params.id;
 
-    User.findByPk(id)
+    const userIdOrder = req.body.userIdOrder;
+
+    User.findByPk(userIdOrder)
         .then(user => {
-            if (user.id == userId) {
-                let image
-                if (req.file) {
-                    image = `${req.protocol}://${req.get('host')}/images/${req.file.filename}`
-                } else {
-                    image == null;
-                }; 
-                if (image == null) {
-                    res.status(400).json({ error: 'Vous n\'avez pas chargé l\'image !' })
-                    console.log(image)
-                } else {
-                    const user = {
-                        imageUrl: image,
-                    };
-                    User.update(user, { where: { id: id } } )
-                        .then(() => res.status(200).json({ message: 'Mot de passe modifié!' }))
-                        .catch((error) => res.status(400).json({ error } ));
-                }
+            if(user.isAdmin == true && user.id == userId) {
+                User.findByPk(id)
+                    .then(user => {
+                        let image
+                        if (req.file) {
+                            image = `${req.protocol}://${req.get('host')}/images/${req.file.filename}`
+                        } else {
+                            image == null;
+                        }; 
+                        if (image == null) {
+                            res.status(400).json({ error: 'Vous n\'avez pas chargé l\'image !' })
+                        } else {
+                            if (user.imageUrl != null) {
+                                const filename = user.imageUrl.split('/images/')[1];
+                                //Utilisation de la fonction unlink pour supprimer l'image et suppression de toute la Sauce
+                                fs.unlink(`images/${filename}`, () => {
+                                const user = {
+                                    imageUrl: image,
+                                };
+                                User.update(user, { where: { id: id } } )
+                                    .then(() => res.status(200).json({ message: 'Photo modifiée en tant qu\'administrateur!' }))
+                                    .catch((error) => res.status(400).json({ error } ));
+                                });
+                            } else {
+                                const user = {
+                                    imageUrl: image,
+                                };
+                                User.update(user, { where: { id: id } } )
+                                    .then(() => res.status(200).json({ message: 'Photo modifiée en tant qu\'administrateur!' }))
+                                    .catch((error) => res.status(400).json({ error } ));
+                            }   
+                        }
+                    })
+                    .catch((error) => res.status(500).json({ error }));
             } else {
-                res.status(401).json({ message: 'Opération non autorisée' } )
-            
+                User.findByPk(id)
+                    .then(user => {
+                        if (user.id == userId) {
+                            let image
+                            if (req.file) {
+                                image = `${req.protocol}://${req.get('host')}/images/${req.file.filename}`
+                            } else {
+                                image == null;
+                            }; 
+                            if (image == null) {
+                                res.status(400).json({ error: 'Vous n\'avez pas chargé l\'image !' })
+                            } else {
+                                if (user.imageUrl != null) {
+                                    const filename = user.imageUrl.split('/images/')[1];
+                                    //Utilisation de la fonction unlink pour supprimer l'image et suppression de toute la Sauce
+                                    fs.unlink(`images/${filename}`, () => {
+                                    const user = {
+                                        imageUrl: image,
+                                    };
+                                    User.update(user, { where: { id: id } } )
+                                        .then(() => res.status(200).json({ message: 'Photo modifiée !' }))
+                                        .catch((error) => res.status(400).json({ error } ));
+                                    });
+                                } else {
+                                    const user = {
+                                        imageUrl: image,
+                                    };
+                                    User.update(user, { where: { id: id } } )
+                                        .then(() => res.status(200).json({ message: 'Photo modifiée !' }))
+                                        .catch((error) => res.status(400).json({ error } ));
+                                }   
+                            }
+                        } else {
+                            res.status(401).json({ message: 'Opération non autorisée' } )
+                        }
+                    })
+                    .catch((error) => res.status(500).json({ error }));
             }
         })
-        .catch((error) => res.status(500).json({ error }));
-
-}
+};
 
 exports.deleteUser = (req, res) => {
     const token = req.headers.authorization.split(' ')[1];
@@ -167,17 +217,52 @@ exports.deleteUser = (req, res) => {
 
     const id = req.params.id;
 
-    User.findByPk(id)
+    const userIdOrder = req.body.userIdOrder;
+
+    User.findByPk(userIdOrder)
         .then(user => {
-            if (user.id == userId) {
-                User.destroy( { where: { id: id } } )
-                    .then(() => res.status(200).json({ message: 'Utilisateur supprimé!'}))
-                    .catch((error) => res.status(400).json({ error }));
+            if(user.isAdmin == true && user.id == userId) {
+                User.findByPk(id)
+                    .then(user => {
+                        if(user.imageUrl != null) {
+                            const filename = user.imageUrl.split('/images/')[1];
+                                //Utilisation de la fonction unlink pour supprimer l'image et suppression de toute la Sauce
+                            fs.unlink(`images/${filename}`, () => {
+                            User.destroy( { where: { id: id } } )
+                                .then(() => res.status(200).json({ message: 'Utilisateur supprimé!'}))
+                                .catch((error) => res.status(400).json({ error }));
+                                });
+                        } else {
+                            User.destroy( { where: { id: id } } )
+                            .then(() => res.status(200).json({ message: 'Utilisateur supprimé!'}))
+                            .catch((error) => res.status(400).json({ error }));
+                            }
+                    })
+                    .catch((error) => res.status(500).json({ error }));
             } else {
-                res.status(401).json({ message: 'Opération non autorisée' })
+                User.findByPk(id)
+                    .then(user => {
+                        if (user.id == userId) {
+                            if(user.imageUrl != null) {
+                            const filename = user.imageUrl.split('/images/')[1];
+                                //Utilisation de la fonction unlink pour supprimer l'image et suppression de toute la Sauce
+                            fs.unlink(`images/${filename}`, () => {
+                            User.destroy( { where: { id: id } } )
+                                .then(() => res.status(200).json({ message: 'Utilisateur supprimé!'}))
+                                .catch((error) => res.status(400).json({ error }));
+                                });
+                            } else {
+                                User.destroy( { where: { id: id } } )
+                                .then(() => res.status(200).json({ message: 'Utilisateur supprimé!'}))
+                                .catch((error) => res.status(400).json({ error }));
+                            }
+                        } else {
+                            res.status(401).json({ message: 'Opération non autorisée' })
+                        }
+                    })
+                    .catch((error) => res.status(500).json({ error }));
             }
         })
-        .catch((error) => res.status(500).json({ error }));
 };
 
 exports.getAllUsers = (req, res) => {
